requirements.txt
pip==24.1.2
setuptools==70.3.0
wheel==0.43.0
requests~=2.32.3
websocket-client~=1.8.0